initial setup - works but need to update search no results
