+++ packages/portal-proto/src/features/cDave/Controls/utils.ts
@@ -1,0 +1,20 @@
+import { sortBy } from "lodash";
+import { FACET_SORT } from "../constants";
+
+export interface CDaveField {
+  readonly field_type: string;
+  readonly field_name: string;
+  readonly description?: string;
+  readonly full: string;
+}
+
+export const sortFacetFields = (
+  fields: CDaveField[],
+  facet_type: string,
+): CDaveField[] => {
+  return sortBy(fields, (item) =>
+    FACET_SORT[facet_type].indexOf(item.field_name) !== -1
+      ? FACET_SORT[facet_type].indexOf(item.field_name)
+      : fields.length,
+  );
+};
